moved annotations definition to xml transformer & merged the two territory filters
