+++ packages/salesforce-adapter/src/filters/territory.ts
@@ -13,10 +13,17 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import { Element, isInstanceElement } from '@salto-io/adapter-api'
+import _ from 'lodash'
+import { Element, isInstanceElement, InstanceElement, getChangeElement } from '@salto-io/adapter-api'
 import { FilterCreator } from '../filter'
-import { isMetadataObjectType, metadataType } from '../transformers/transformer'
-import { TERRITORY2_TYPE, TERRITORY2_MODEL_TYPE } from '../constants'
+import { isMetadataObjectType, metadataType, apiName } from '../transformers/transformer'
+import { NESTED_DIR_NAMES_ANNOTATION, CONTENT_FILENAME_ANNOTATION } from '../transformers/xml_transformer'
+import { TERRITORY2_TYPE, TERRITORY2_MODEL_TYPE, TERRITORY2_RULE_TYPE } from '../constants'
+
+const territory2TypesToNestedDirName: Record<string, string> = {
+  [TERRITORY2_RULE_TYPE]: 'rules',
+  [TERRITORY2_TYPE]: 'territories',
+}
 
 const removeCustomFieldsFromTypes = (elements: Element[], typeNames: string[]): void => {
   const elementsOfTypes = elements.filter(elem => typeNames.includes(metadataType(elem)))
@@ -32,12 +39,50 @@ const removeCustomFieldsFromTypes = (elements: Element[], typeNames: string[]): 
     })
 }
 
+const contentFileNameAnnotate = (elements: InstanceElement[], typeNames: string[]): void => {
+  const elementsOfTypes = elements.filter(elem => typeNames.includes(metadataType(elem)))
+  elementsOfTypes.forEach(element => {
+    const nameWithoutModelPrefix = apiName(element).split('.').slice(1).join('.')
+    const { suffix } = element.type.annotations
+    const contentFileName = `${nameWithoutModelPrefix}${suffix === undefined ? '' : `.${suffix}`}`
+    element.annotate({ [CONTENT_FILENAME_ANNOTATION]: contentFileName })
+  })
+}
+
+const nestedDirsAnnotate = (elements: InstanceElement[], typeNames: string[]): void => {
+  const elementsOfTypes = elements.filter(elem => typeNames.includes(metadataType(elem)))
+  elementsOfTypes.forEach(element => {
+    const modelName = apiName(element).split('.')[0]
+    const nestedDirsNames = [modelName]
+
+    if (territory2TypesToNestedDirName[metadataType(element)] !== undefined) {
+      const nestedDirName = territory2TypesToNestedDirName[metadataType(element)]
+      nestedDirsNames.push(nestedDirName)
+    }
+    element.annotate({ [NESTED_DIR_NAMES_ANNOTATION]: nestedDirsNames })
+  })
+}
+
 const filterCreator: FilterCreator = () => ({
   onFetch: async elements => {
     // Territory2 and Territory2Model support custom fields - these are returned
     // in a CustomObject with the appropriate name and also in each instance of these types
     // We remove the fields from the instances to avoid duplication
     removeCustomFieldsFromTypes(elements, [TERRITORY2_TYPE, TERRITORY2_MODEL_TYPE])
+  },
+  // territory2 types require a special deploy pkg structure (SALTO-1200)
+  preDeploy: async changes => {
+    const instanceElements = changes.map(getChangeElement).filter(isInstanceElement)
+    contentFileNameAnnotate(instanceElements, Object.keys(territory2TypesToNestedDirName))
+    nestedDirsAnnotate(instanceElements,
+      [...Object.keys(territory2TypesToNestedDirName), TERRITORY2_MODEL_TYPE])
+  },
+  onDeploy: async changes => {
+    changes.map(getChangeElement).forEach(elem => {
+      elem.annotations = _.omit(elem.annotations,
+        [NESTED_DIR_NAMES_ANNOTATION, CONTENT_FILENAME_ANNOTATION])
+    })
+    return []
   },
 })
